+++ include/vnx/TerminalInput.h
@@ -32,8 +32,8 @@ public:
 	void add_argument(const std::string &command, const std::string &argument);
 	void clear_commands();
 	void clear_arguments(const std::string &command);
-	std::string complete_command(const std::string &prefix);
-	std::string complete_argument(const std::string &command, const std::string &prefix);
+	std::pair<std::string, bool> complete_command(const std::string &prefix);
+	std::pair<std::string, bool> complete_argument(const std::string &command, const std::string &prefix);
 	std::vector<std::string> guess_command(const std::string &prefix);
 	std::vector<std::string> guess_argument(const std::string &command, const std::string &prefix);
 
@@ -43,7 +43,7 @@ private:
 	std::list<std::string>::iterator m_historyPos;
 	std::set<std::string> m_commands;
 	std::map<std::string, std::set<std::string>> m_arguments;
-	std::string complete(const std::string &prefix, const std::set<std::string> &set);
+	std::pair<std::string, bool> complete(const std::string &prefix, const std::set<std::string> &set);
 	std::vector<std::string> guess(const std::string &prefix, const std::set<std::string> &set);
 };